Moved e2e testings into seperate dir e2e
