+++ server/src/models/Friend.ts
