+++ server/src/models/Scalar.ts
