+++ server/src/resolvers/Friend/query.ts
@@ -1,0 +1,2 @@
+import { getUserId } from '../../utils/auth';
+import { queryField } from '@nexus/schema';
