--- server/src/types/resolvers/Friend/query.ts
@@ -1,2 +1,0 @@
-import { getUserId } from '../../../utils/auth';
-import { queryField } from '@nexus/schema';
