+++ server/tests/e2e/resolvers/User/mutation/signInWithFacebook.test.ts
@@ -1,5 +1,6 @@
+import * as AuthUtils from '../../../../../src/utils/auth';
+
 import { request } from 'graphql-request';
-import * as AuthUtils from '../../../../src/utils/auth';
 import { signInWithFacebook } from '../../../setup/queries';
 import { testHost } from '../../../setup/testSetup';
