+++ server/tests/e2e/resolvers/friend.test.ts
